Merged PR 7099: Add IdentityCoreFork to the project
## Proposed changes

This PR updates the url of the submodule `IdentityCore` to a forked version of it.

This is the fork: https://identitydivision.visualstudio.com/DefaultCollection/Engineering/_git/devexdub-microsoft-authentication-library-common-for-objc

The submodule's default branch for us will be `ciam-dev`. The idea is to have `dev` as a clean branch (we will just pull the latest changes from the MSAL team on it).

## Important

To make these changes happen in your local repo:

1. Open the command line tool and run `git submodule sync`. This will update the origin of the submodule.
2. Run `git submodule update --remote` to make sure you pull the latest changes from the submodule.
3. Check in SourceTree (or whatever UI program you're using) that you're the changes have taken place (in SourceTree you can see a "View Remote" button that launches the repo in your browser).

Related work items: #2183650