Merged PR 7136: Define components internal interface
## Proposed changes

This PR adds a basic interface for the following components:
- Logger
- Cache
- Telemetry

This internal API is meant to be used by us during development, we will be using this code from our Swift classes and the implementation of these protocols will interact with obj-c code, so feel free to suggest anything you'd like to have.

Note: It's not a final API since we still don't know some things, such as the final telemetry requirements, etc.

Related work items: #2176394