Merged PR 7211: Initial baseline Native Auth SDK sample app
## Proposed changes

This PR adds the initial sample app for the Native Auth functionality to the MSAL repo. The sample app is based on the POC app created in Q3, with updates to include the MSAL framework directly as a subproject, formatting and coding style updates, the addition of SwiftLint as a step during compilation, etc.

## Type of change

- [X] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

The sample app is not functional at this point and does not make any calls into the MSAL library for new or old functionality.  Furthermore, the signature of the `IClientAuthSDKCommunicator` protocol remains as it was for the Q3 POC. Updating the signature and connecting this app to the Native Auth functionality of MSAL is not part of this pull request.

Related work items: #2126798, #2126812