Merged PR 7267: Groundwork Cache component implementation
## Proposed changes

In this PR I added a new Gateway class to interact with the existing MSAL caching component. I spent most of the time to understand the configuration needed to save and retrieve tokens and account. You can find some examples in the unit test.
I didn't use the "Handler" and the "Validator" classes because I think that they belong more to the Network component than the Caching one.

## Type of change

- [X] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [X] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

Code coverage of the new class:
![Screenshot 2022-12-30 at 15.43.48.png](https://identitydivision.visualstudio.com/fac9d424-53d2-45c0-91b5-ef6ba7a6bf26/_apis/git/repositories/616da9d3-8336-4562-811a-581d9a5cbe9c/pullRequests/7267/attachments/Screenshot%202022-12-30%20at%2015.43.48.png)

Related work items: #2139086