Merged PR 7296: Adding CI build support for CIAM sample app
## Proposed changes

Updated the existing MSAL build script and PR validation Azure Pipeline definition to add support for building and running SwiftLint on the CIAM sample app.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [X] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [X] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

Once this PR is merged to ciam-master we can configure an Azure pipeline to automatically build, lint and test the sample app when a PR is created and when changes are merged to a main branch.

I went with the option of adding to the existing MSAL CI process and build script rather than starting from scratch as it will help when we merge with MSAL on GitHub. The downside to this is that there is limited output in Azure Dev Ops about failures in the build and lint processes.

This PR also updates the Sample App Xcode project to add a missing dependency on the MSAL framework that is needed to allow command line build to succeed.