Merged PR 7298: Add SwiftLint to MSAL Xcode project
## Proposed changes

This PR adds a call to the SwiftLint tool during the build of the MSAL iOS Framework.

The linter only runs on Swift code within the MSAL subdirectory so Swift code in Tests folder will not be linted.

I added some line length exceptions to the SwiftLint configuration to avoid having to make too many changes to the existing code, but otherwise the rules are the default SwiftLint rules.  I made changes the existing code to satisfy the rules. These changes were to the formatting of the code only - no functional changes to the existing code were made as part of this PR.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [X] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [X] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

- If SwiftLint is not installed on a developer's machine, the script will issue a warning and continue with the build.
- The SwiftLint step will be skipped if the build is running on a Azure DevOps pipeline build server

## Proposed changes

Describe what this PR is trying to do.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information