Merged PR 7538: Adding new Native Auth SDK sample app
## Proposed changes

This PR adds the initial new Native Auth SDK sample app as described in [this doc](https://microsofteur.sharepoint.com/:w:/t/DevExDublin/EQF7atYNfCZDuaqXaesDyEQBnpjfwFxHCAZokz19DDFS8g?e=oORFvR).

This initial PR is just the skeleton of the app with no real code.

Also added to the build script so it will be built and linted with PRs.

The existing "legacy" sample app remains in place for now. We may update this for internal use to test out scenarios.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [X] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [X] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

Related work items: #2398485