Merged PR 7571: Add Local Telemetry handling to controllers
## Proposed changes

This PR adds telemetry handling to the controllers. I've used the same approach as MSAL: creating a base controller from which the controllers will inherit.

I pondered about using a protocol approach, but I saw in `MSIDLocalInteractiveController` that one of the methods was overridden, and it was extending the parent implementation, so I though maybe in the future it could be useful to use inheritance here.

From now on, all the controllers will need to inherit from the `MSALNativeAuthBaseController`.

This PR needs this code to work properly: !7609

## Type of change

- [x] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [x] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

Related work items: #2411852