Merged PR 7583: Add Sign-Up
## Proposed changes

This PR adds the sign up flow.

The code could be thinner if we merged some of the SignUp + SignIn code, as they share many common things. However, as we are planning to update this code with Mock API v2, I think it's better to have separate flows for now and make the decision when we adapt to the new changes of the API.

If you want to test it, the same comment from the [SignIn PR](https://identitydivision.visualstudio.com/Engineering/_git/devexdub-microsoft-authentication-library-for-objc/pullrequest/7448) applies:

Our current backend response doesn't include home_account_id, so in order to test it with the mock api we need to mock in it IdentityCore:

In MSIDAADOauth2Factory.m:

```
- (MSIDAccountIdentifier *)accountIdentifierFromResponse:(MSIDAADTokenResponse *)response
{
    return [[MSIDAccountIdentifier alloc] initWithDisplayableId:response.idTokenObj.username
                                                  homeAccountId:@"mock-home-account"];
}
```

## Type of change

- [x] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [x] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

Related work items: #2283285, #2283288