Merged PR 7510: Verify Code
Our current backend response doesn't include home_account_id, so in order to test it with the mock api we need to mock in it IdentityCore:

In MSIDAADOauth2Factory.m:

```
- (MSIDAccountIdentifier *)accountIdentifierFromResponse:(MSIDAADTokenResponse *)response
{
    return [[MSIDAccountIdentifier alloc] initWithDisplayableId:response.idTokenObj.username
                                                  homeAccountId:@"mock-home-account"];
}
```

## Proposed changes

Added MSALNativeAuthVerifyCodeRequest and helper classes that are used to make the request to the server.
Added MSALNativeAuthVerifyCodeController which is used by MSALNativePublicClientApplication to initiate the request
Added method to MSALNativeAuthRequestProvider to generate the MSALNativeAuthVerifyCodeRequest
Added MSALNativeAuthVerifyCodeResponse class to parse response from server

## Type of change

- [x] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [x] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information
Unit tests are also present!

Related work items: #2328400