Merged PR 8145: Update CI build pipeline to use macOS 12
## Proposed changes

This PR fixes the broken build by using the macOS-12 image instead of macOS-11. This is a deviation from the MSAL official pipelines so we may need to revert at a later date when merging to MSAL ObjC main branch.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [X] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

The MSAL CI builds are currently broken, getting stuck running tests. After several days of troubleshooting it is unclear what the underlying issue is. This quick upgrade fixes our build and allows building, linting and testing for PR validation.