Merged PR 8374: Add initialiser for Objective-C (plus changes to sample app to validate it)
## Proposed changes

This PR replaces the `MSALNativeAuthChallengeType` `@objc enum` and replaces it with an [OptionSet](https://nshipster.com/optionset/.) called `MSALNativeAuthChallengeTypes` which is defined in Objective-C to allow for best interoperability (see Additional Information below).

The existing intialiser `init(configuration config: MSALPublicClientApplicationConfig, challengeTypes: [MSALNativeAuthChallengeType]) throws` is not accessible from Objective-C because it includes an array of an enum.

However, now that MSALNativeAuthChallengeTypes is an OptionSet, it means that the same initialiser is accessible from both Swift and Objective-C.

When called from Objective-C the method looks like this:

```objc
    self.nativeAuth = [[MSALNativeAuthPublicClientApplication alloc]
                       initWithConfiguration:configuration
                       challengeTypes:MSALNativeAuthChallengeTypeOOB | MSALNativeAuthChallengeTypePassword
                       error:&error];
```

When called from Swift it looks like this (for a single challenge type):

```swift
  nativeAuth = try MSALNativeAuthPublicClientApplication(
                configuration: config,
                challengeTypes: .OOB
            )
```

or this, when there are several challenge types:

```swift
  nativeAuth = try MSALNativeAuthPublicClientApplication(
                configuration: config,
                challengeTypes: [.OOB, .password]
            )
```

Note: Previously we had been using `oob` as the challenge type name. Due to Objective-C -> Swift conversion, this is now `OOB`.

This PR also includes updates to the Native Auth sample app to show the new initialiser in action.

If OptionSets or bitmasks are new to you, check out this NSHipster article that I also linked above: https://nshipster.com/optionset/.

## Type of change

- [ ] Feature work
- [X] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [X] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

If the new OptionSetType was defined in Swift then we would have to implement it as an `@objc class` instead of a `struct` and then had to add various bits of fragile looking boilerplate to make it work (see https://stackoverflow.com/a/52207154/2347353).

Related work items: #2571162