Merged PR 8466: Add sign-up controller v2
## Proposed changes

This PR adds the sign-up controller working with the mock API v2.

I've removed the legacy classes because I was starting to have issues with them. @<Danilo Raspa>, as you were going to remove them as well, if you have huge conflicts, maybe one of us can open a PR to remove those classes and we both can rebase on that one. That approach may be cleaner.

Note 1: I'm using `async/await` because after talking to Danilo, he said we could use it as long as we don't propagate the `async` word to the public interface.

Note 2: I haven't created an intermediate error model (similar to `MSALNativeAuthSignInTokenValidatedErrorType`). Instead, I return the api errors to the controller, and they're converted into the public errors with a mapping function. I think having an extra "validated error model" is not really needed and adds complexity. But I'm open to discuss this with the team.

## Type of change

- [x] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [x] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

Related work items: #2457579