Remove "invalid_attributes" error from signup/start and signup/continue errors as it has now become "attribute_validation_failed"
Remove check for signUpToken in validator for certain errors according to the doc API Response Validation. Changes in SignUpValidatedResponses + SignUpController.
Add signUpToken to the MSALNativeAuthCacheAccessor. Added some tests in SignUpControllerTests that check for cached signUpToken

Add tests for CacheAccessor handling the signUpToken

Add tests for the parameters that the Controller sends to the RequestProvider

Remove legacy code

Refactor attribute_validation_failed errors in /start and /continue

Replace error with AttributeRequiredState

Use signUpToken internally