Merged PR 8994: Make all sign-in flows start from /initiate
## Proposed changes

The main changes are:

- Remove nca flag
- Optional implementation of `onSignInCodeRequired()` has been added to `SignInPasswordStartDelegate`.
- SignIn with password now uses /initiate + /challenge + /token.
    - If the /challenge endpoint returns `challengeType = OOB`, we log a warning to the user and call `delegate.onSignInCodeRequired()`.
- Removed `.invalidAuthenticationType` error from /token endpoint.
- Sample app NOT updated - @<Rodhan Hickey> let me know if you think we should update it.

I've created a tuple at the BaseController level `typealias TelemetryInfo = (event?, context)`, since in order to handle the telemetry event we also need the context, this helps to have one parameter less in most of the private functions. Let me know if you don't like it and I will move back to having the event + context separate. If you like it we can extend it to the rest of controllers in a separate PR.

I haven't migrated to the use of `@MainActor` because the Controller already has quite a few changes, but I can create a small PR later. Preferably once both this PR and the cache fix PR are merged (to avoid even more conflicts in the sign-in controller).

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [x] Engineering change
- [ ] Test
- [ ] Logging/Telemetry

## Risk

- [x] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

Related work items: #2616683