Merged PR 9103: Fix sign out e2e tests for mock API, skip for test tenant
## Proposed changes

Small PR to fix sign out e2e tests for mock API, skip for test tenant.

Now that we no longer go straight to the token endpoint when signing in with a password, we need to provide queue up the mock responses for the /initiate and /challenge endpoints too.

Also, as this flow isn't supported in the test tenant at the moment we need to skip running this if it is not a Mock API test configuration.

## Type of change

- [ ] Feature work
- [ ] Bug fix
- [ ] Documentation
- [ ] Engineering change
- [X] Test
- [ ] Logging/Telemetry

## Risk

- [ ] High – Errors could cause MAJOR regression of many scenarios. (Example: new large features or high level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios. (Example: somewhat complex bug fixes, small new features)
- [X] Small – No issues are expected. (Example: Very small bug fixes, string changes, or configuration settings changes)

## Additional information

Fix sign out e2e tests for mock API, skip for test tenant